+++ i3-save.sh
@@ -2,6 +2,10 @@
 
 set -e
 IFS=$'\n\t'
+CURR_DIR=$(dirname "${0}")
+
+# Handle errors
+source "${CURR_DIR}/utils/error_handling.sh"
 
 # Check for the version flag in every argument
 if [[ ! "${@#--version}" = "$@" || ! "${@#-v}" = "$@" ]]; then
@@ -12,15 +16,19 @@ fi
 
 # Check if user has jq installed
 if ! command -v jq >/dev/null 2>&1; then
-	echo "jq is required for i3-restore!"
+	error "jq is required for i3-restore!"
 	exit
 fi
 
 # Check for perl-anyevent-i3 by seeing if i3-save-tree fails
 if [[ $(i3-save-tree 2>&1) == "Can't locate AnyEvent/I3.pm"* ]]; then
-	echo "perl-anyevent-i3 is required for i3-restore!"
+	error "perl-anyevent-i3 is required for i3-restore!"
 	exit
 fi
+
+# Start logger and import log function
+source "${CURR_DIR}/utils/logs.sh"
+rotate_log
 
 # Set default if not configured
 i3_PATH="${i3_PATH:=${HOME}/.config/i3}"
@@ -31,7 +39,10 @@ workspaces=$(echo "${ALLWORKSPACES}" | jq -r '.[] | .name') # Match only the wor
 # Remove previous saved session. -f is there so it doesn't exit if rm fails (file isn't found)
 rm -f "${i3_PATH}"/*.json
 
+log "Saving current i3wm session"
+
 for ws in ${workspaces}; do
+    log "Saving layout for Workspace ${ws}"
     # Replace slash in workspace name as file names cannot have slashes
     sanitized_ws_name=${ws//\//\{slash\}}
 
@@ -40,6 +51,7 @@ for ws in ${workspaces}; do
 
     # If workspace is empty (i.e doesn't contain any actual configuration lines)
     if [[ ! $workspace_tree == *"{"* ]]; then
+        log "Empty layout for Workspace ${ws}. Skipping..."
         continue
     fi
 
@@ -49,9 +61,12 @@ for ws in ${workspaces}; do
     sed -i 's|^\(\s*\)// "|\1"|g; /^\s*\/\//d' "$file_name"
 done
 
-DIR=$(dirname "${0}")
-python "${DIR}"/programs/i3-save.py
+error_message=$(python "${CURR_DIR}/programs/i3-save.py")
+[[ ! -z "${error_message}" ]] && error "An error occured saving the session's programs. View the logs for more details" 1
 
-# Execute command passed as an argument. For use with the i3 config file and so it can be manually executed
+log "Finished saving current i3wm session"
+
+# Execute commands passed as an argument. For use with the i3 config file and so it can be manually executed
 # without exiting out of the session
+[ "$@" ] && log "Executing arguments as a command: $@" # Only log if arguments are passed in
 "$@"