+++ programs/utils.py
@@ -1,5 +1,10 @@
 import json
+import logging
+import os
 import subprocess
+import sys
+
+DEFAULT_LOG_FILE = "logs/i3-restore.log"
 
 
 def get_workspaces():
@@ -24,3 +29,18 @@ def get_workspaces():
 def get_tree():
     tree = subprocess.check_output(["i3-msg", "-t", "get_tree"]).decode("utf-8")
     return json.loads(tree)
+
+
+def get_logger():
+    project_dir = os.path.dirname(os.path.dirname(__file__))
+    log_file = os.getenv("I3_RESTORE_LOG_FILE", f"{project_dir}/{DEFAULT_LOG_FILE}")
+
+    logging.basicConfig(filename=log_file, filemode="a", level=logging.DEBUG, format="")
+    logger = logging.getLogger()
+
+    # Print error messages to stdout
+    handler = logging.StreamHandler(sys.stdout)
+    handler.setLevel(logging.ERROR)
+    logger.addHandler(handler)
+
+    return logger