+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 When upgrading to a new version, make sure to follow the directions under the "Upgrading" header of the corresponding version.
 If there is no "Upgrading" header for that version, no post-upgrade actions need to be performed.
 
+## 1.2 (2022-10-29)
+
+### New Features
+- A logger was added to make it easier to debug the script
+- An [i3-nagbar](https://man.archlinux.org/man/community/i3-wm/i3-nagbar.1.en) will now appear when an error occurs while
+running the script, making it easier for the user to see what is wrong.
+
+### Bug Fixes
+- Subprocess programs are now saved correctly (e.g vim, cmus, etc.)
+
+
 ## 1.1 (2022-08-13)
 
 ### New Features